Move Context to its own file
